feat: implement unitary and binary operations in JITCMatrix class
